requirements.txt
aiocfscrape==1.0.0
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioquic==1.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
cachetools==4.2.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
cryptography==44.0.0
deepdiff==8.1.1
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huepy==1.2.1
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Js2Py==0.74
kaitaistruct==0.10
ldap3==2.9.1
magic-filter==1.0.12
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
orderly-set==5.2.3
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
propcache==0.2.1
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydivert==2.1.0
pyjsparser==2.7.1
pylsqpack==0.3.18
pyOpenSSL==24.3.0
pyparsing==3.2.1
pyperclip==1.9.0
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
retry-requests==2.0.0
ruamel.yaml==0.18.10
selenium==4.28.1
selenium-stealth==1.0.6
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
tornado==6.4.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
urwid==2.6.16
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
